+++ src/main/java/com/jj/Gradebook/dao/AttendanceRepository.java
@@ -3,5 +3,5 @@ package com.jj.Gradebook.dao;
 import com.jj.Gradebook.entity.Attendance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttendanceRepository extends JpaRepository<Attendance,Integer> {
+public interface AttendanceRepository extends JpaRepository<Attendance,Long> {
 }
