+++ src/main/java/com/jj/Gradebook/dao/CourseRepository.java
@@ -3,5 +3,5 @@ package com.jj.Gradebook.dao;
 import com.jj.Gradebook.entity.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseRepository extends JpaRepository<Course, Integer> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
 }
