+++ src/main/java/com/jj/Gradebook/service/course/CourseService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface CourseService {
     List<Course> findAll();
-    Course findById(int id);
+    Course findById(Long id);
     Course save(Course course);
-    void deleteById(int id);
+    void deleteById(Long id);
 }
